backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.14.1
python-dotenv==1.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.3
flask-mail==0.9.1
cloudinary==1.36.0
waitress==3.0.2
firebase-admin==6.4.0
